add --existing option
This allows the user to pass in the filename of an existing
JSON file with an "ids" key. Since Twitter both limits the number
of tweets that can be retrieved in one request and rate limits
requests, this option is necessary for collecting tweet IDs of
users with many thousands of tweets.